Added unique indexes for user table, changed type of id for a user entity, Added plane table, repository and service.

+++ Controllers/UserController.cs
@@ -19,7 +19,7 @@ public class UserController : ControllerBase
 
 	[HttpGet]
 	[Route("{id}")]
-	public ActionResult<User> Get([FromRoute] int id)
+	public ActionResult<User> Get([FromRoute] long id)
 	{
 		var user = _userService.GetById(id);
 		return user;
